split mqtt window into separate module
