+++ src/lib.rs
@@ -27,6 +27,8 @@ pub fn get_app_version() -> &'static Version {
     })
 }
 
+#[cfg(feature = "mqtt")]
+pub mod mqtt;
 pub mod plot;
 #[cfg(all(feature = "profiling", not(target_arch = "wasm32")))]
 pub mod profiling;
